requirements.txt
appnope==0.1.0
backcall==0.1.0
blinker==1.4
click==6.7
decorator==4.3.0
Flask==1.0.2
Flask-DebugToolbar==0.10.1
Flask-Modus==0.0.1
Flask-SQLAlchemy==2.3.2
ipython==6.4.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
MarkupSafe==1.0
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2-binary==2.7.5
ptyprocess==0.5.2
Pygments==2.2.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.8
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
